bluetooth/requestLEScan/doesnt-consume-user-gesture.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
const test_desc = 'requestLEScan calls do not consume user gestures.';

if (!navigator.bluetooth.requestLEScan) {
  // requestLEScan is being tested behind a flag. When we get further along in the
  // implementation, we will remove this if-statement.
  done();
}
bluetooth_test(() => navigator.bluetooth.test.simulateCentral({ state: 'powered-on' })
      .then(() => callWithTrustedClick(() => {
      let first = navigator.bluetooth.requestLEScan();
      let second = navigator.bluetooth.requestLEScan();
      return Promise.all([
        first.then(scan => assert_equals(
          scan.constructor.name, 'BluetoothLEScan')),
        second.then(scan => assert_equals(
          scan.constructor.name, 'BluetoothLEScan')),
      ]);
    })), test_desc);
</script>
